Mavenize - reorganize files
