--- ivy.xml
@@ -1,36 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE ivy-module [
-	<!ENTITY vaadin.version "7.1.1">
-]>
-<ivy-module version="2.0"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:noNamespaceSchemaLocation="http://ant.apache.org/ivy/schemas/ivy.xsd">
-	<info organisation="com.example" module="v7proj" />
-	<configurations>
-		<!-- The default configuration, which should be deployed to the server -->
-		<conf name="default" />
-		<!-- A configuration only needed when compiling the widget set. Should 
-			not be deployed to the server -->
-		<conf name="widgetset-compile" />
-	</configurations>
-	<dependencies>
-		<!-- The core server part of Vaadin -->
-		<dependency org="com.vaadin" name="vaadin-server" rev="&vaadin.version;"
-			conf="default->default" />
-
-		<!-- Vaadin themes -->
-		<dependency org="com.vaadin" name="vaadin-themes" rev="&vaadin.version;" />
-
-		<!-- Precompiled DefaultWidgetSet -->
-		<dependency org="com.vaadin" name="vaadin-client-compiled"
-			rev="&vaadin.version;" />
-
-		<!-- Vaadin client side, needed for widget set compilation -->
-		<dependency org="com.vaadin" name="vaadin-client" rev="&vaadin.version;" />
-
-		<!-- Compiler for custom widget sets. Should not be deployed -->
-		<dependency org="com.vaadin" name="vaadin-client-compiler"
-			rev="&vaadin.version;" conf="widgetset-compile->default" />
-	</dependencies>
-</ivy-module>
-
