--- popupextension-addon/src/main/java/org/vaadin/addons/client/MyComponentServerRpc.java
@@ -1,12 +1,0 @@
-package org.vaadin.addons.client;
-
-import com.vaadin.shared.MouseEventDetails;
-import com.vaadin.shared.communication.ServerRpc;
-
-// ServerRpc is used to pass events from client to server
-public interface MyComponentServerRpc extends ServerRpc {
-
-    // Example API: Widget click is clicked
-    public void clicked(MouseEventDetails mouseDetails);
-
-}
