--- popupextension-addon/src/main/java/org/vaadin/addons/client/MyComponentState.java
@@ -1,8 +1,0 @@
-package org.vaadin.addons.client;
-
-public class MyComponentState extends com.vaadin.shared.AbstractComponentState {
-
-    // State can have both public variable and bean properties
-    public String text = "MyComponent";
-
-}
