--- popupextension-addon/src/main/java/org/vaadin/addons/client/MyComponentWidget.java
@@ -1,16 +1,0 @@
-package org.vaadin.addons.client;
-
-import com.google.gwt.user.client.ui.Label;
-
-// Extend any GWT Widget
-public class MyComponentWidget extends Label {
-
-    public MyComponentWidget() {
-
-        // CSS class-name should not be v- prefixed
-        setStyleName("popupextension");
-
-        // State is set to widget in MyComponentConnector
-    }
-
-}
