+++ popupextension-addon/src/main/java/org/vaadin/addons/client/PopupExtensionDataTransferComponentState.java
@@ -1,0 +1,9 @@
+package org.vaadin.addons.client;
+
+import com.vaadin.shared.AbstractComponentState;
+
+public class PopupExtensionDataTransferComponentState extends
+        AbstractComponentState {
+    private static final long serialVersionUID = 6235786382671700479L;
+    public String popupId;
+}
