+++ popupextension-addon/src/main/java/org/vaadin/addons/client/PopupExtensionDataTransferComponentWidget.java
@@ -1,10 +1,10 @@
-package com.github.wolfie.popupextension.client;
+package org.vaadin.addons.client;
 
 import com.google.gwt.user.client.DOM;
 import com.google.gwt.user.client.ui.Widget;
 
 public class PopupExtensionDataTransferComponentWidget extends Widget {
-	public PopupExtensionDataTransferComponentWidget() {
-		setElement(DOM.createDiv());
-	}
+    public PopupExtensionDataTransferComponentWidget() {
+        setElement(DOM.createDiv());
+    }
 }
