+++ popupextension-addon/src/main/java/org/vaadin/addons/client/PopupExtensionServerRpc.java
@@ -1,7 +1,7 @@
-package com.github.wolfie.popupextension.client;
+package org.vaadin.addons.client;
 
 import com.vaadin.shared.communication.ServerRpc;
 
 public interface PopupExtensionServerRpc extends ServerRpc {
-	void setOpen(boolean open);
+    void setOpen(boolean open);
 }
