+++ popupextension-addon/src/main/resources/org/vaadin/addons/PopupextensionWidgetset.gwt.xml
@@ -2,5 +2,5 @@
 <!DOCTYPE module PUBLIC "-//Google Inc.//DTD Google Web Toolkit 1.7.0//EN" "http://google-web-toolkit.googlecode.com/svn/tags/1.7.0/distro-source/core/src/gwt-module.dtd">
 <module>
 	<inherits name="com.vaadin.DefaultWidgetSet" />
-	<stylesheet src="popupextension.css"/>
+	<stylesheet src="popupextension/styles.css"/>
 </module>
