--- popupextension-addon/src/main/resources/org/vaadin/addons/WidgetSet.gwt.xml
@@ -1,11 +1,0 @@
-<module>
-
-	<!-- WS Compiler: manually edited -->
-
-	<!-- Inherit DefaultWidgetSet -->
-	<inherits name="com.vaadin.DefaultWidgetSet" /> 
-
-	<!-- Widget styles in public -directory -->
-	<stylesheet src="popupextension/styles.css"/>
-
-</module>
