--- src/com/github/wolfie/popupextension/public/popupextension.css
@@ -1,7 +1,0 @@
-.h-popupextension {
-	background-color: white;
-	-moz-box-sizing: border-box;
-	-webkit-box-sizing: border-box;
-	box-sizing: border-box;
-	padding: 10px;
-}
